separated tech recyclerview adapter from the list activity class
