+++ app/src/main/java/com/jasminebreedlove/techandsportsnews/tech/TechArticleListActivity.kt
@@ -18,12 +18,13 @@ import org.jetbrains.anko.AnkoLogger
 
 import android.arch.lifecycle.Observer
 import android.arch.lifecycle.ViewModelProviders
+import android.view.View.GONE
+import android.view.View.VISIBLE
 import java.util.*
 
 class TechArticleListActivity : AppCompatActivity(), AnkoLogger {
 
-    private var twoPane: Boolean = false
-    var articles: ArrayList<Article> = ArrayList()
+    private var twoPane: Boolean = false // todo: might need to move to viewmodel
     lateinit var techViewModel: TechViewModel
 
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -44,12 +45,16 @@ class TechArticleListActivity : AppCompatActivity(), AnkoLogger {
     } // onCreate()
 
     fun setupRecyclerView() {
+        progressBarTech.visibility = VISIBLE // todo: figure out why progress bar isn't showing
         // create observer
         val articleListObserver = Observer { articleList: ArrayList<Article>? ->
             if (articleList != null) {
-                article_list.adapter = TechArticleListActivity.ArticleRecyclerViewAdapter(this@TechArticleListActivity, articleList, twoPane)
+                val techAdapter = TechArticleRecycler(this, articleList, twoPane)
+                article_list.adapter = techAdapter
             }
         }
+
+        progressBarTech.visibility = GONE
             // another way to write the observer
 //        val seekBarValObserver = object : Observer<ArrayList<Article>> {
 //            override fun onChanged(t: ArrayList<Article>?) {
@@ -79,70 +84,4 @@ class TechArticleListActivity : AppCompatActivity(), AnkoLogger {
     } // getTechFeed()*/
 
     // todo: add article link to view
-    class ArticleRecyclerViewAdapter(private val parentActivity: TechArticleListActivity,
-                                     private val articles: ArrayList<Article>,
-                                     private val twoPane: Boolean) :
-            RecyclerView.Adapter<ArticleRecyclerViewAdapter.ViewHolder>() {
-
-        private val onClickListener: View.OnClickListener
-
-        init {
-            onClickListener = View.OnClickListener { v ->
-                val article = v.tag as Article
-                if (twoPane) {
-                    val fragment = TechArticleDetailFragment().apply {
-                        arguments = Bundle().apply {
-                            putString(TechArticleDetailFragment.ARTICLE_TITLE, article.articleTitle)
-                            putString(TechArticleDetailFragment.ARTICLE_LINK, article.link)
-                            putString(TechArticleDetailFragment.ARTICLE_PUB, article.pubDate)
-                            putString(TechArticleDetailFragment.ARTICLE_DESCRIPTION, article.description)
-                            putString(TechArticleDetailFragment.ARTICLE_CATEGORY, article.category)
-                        }
-                    }
-                    parentActivity.supportFragmentManager
-                            .beginTransaction()
-                            .replace(R.id.article_detail_container, fragment) // replace existing article content
-                            .commit()
-                } else {
-                    addTechFragment(article, v)
-
-                }
-            }
-        }
-
-        private fun addTechFragment(article: Article, v: View) {
-            val intent = Intent(v.context, TechArticleDetailActivity::class.java).apply {
-                putExtra(TechArticleDetailFragment.ARTICLE_TITLE, article.articleTitle)
-                putExtra(TechArticleDetailFragment.ARTICLE_LINK, article.link)
-                putExtra(TechArticleDetailFragment.ARTICLE_PUB, article.pubDate)
-                putExtra(TechArticleDetailFragment.ARTICLE_DESCRIPTION, article.description)
-                putExtra(TechArticleDetailFragment.ARTICLE_CATEGORY, article.category)
-            }
-            v.context.startActivity(intent)
-        }
-
-
-        override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
-            val view = LayoutInflater.from(parent.context)
-                    .inflate(R.layout.article_list_content, parent, false)
-            return ViewHolder(view)
-        }
-
-        override fun onBindViewHolder(holder: ViewHolder, position: Int) {
-            val article = articles[position]
-            holder.articleTitle.text = article.articleTitle
-
-            with(holder.articleTitle) {
-                tag = article
-                setOnClickListener(onClickListener)
-            }
-        }
-
-        override fun getItemCount() = articles.size
-
-        inner class ViewHolder(view: View) : RecyclerView.ViewHolder(view) {
-            val articleTitle: TextView = view.article_title
-        } // inner custom ViewHolder class
-
-    } // custom RecyclerAdapter class
 }
